Corrected bins

+++ builder/analysis.cxx
@@ -197,11 +197,11 @@ int main(int argc, char *argv[])
 				Jet *jet = (Jet*) branchGenJet->At(i);
 				if (abs(jet->Eta)>ana.maxEta) continue;
 				if (jet->PT      <ana.minPT)  continue;
-                                if (jet->Mass<1)                    continue; // cannot be zero 
+                                if (jet->Mass<1)              continue; // cannot be zero 
 
 
                                 // get fraction of b-quark inside this jet (in %) 
-                                float btag_fracmom=0;
+                                double btag_fracmom=0;
                                 for(int i1 = 0; i1 < branchParticle->GetEntriesFast(); ++i1) {
                                         GenParticle *ph = (GenParticle*) branchParticle->At(i1);
                                         int    pdgCode = TMath::Abs(ph->PID);
